refactor: renamed and moved button component that is part of the TokensMenu list (#1522)
