+++ src/frontend/src/icp-eth/components/tokens/ManageTokensMenuButton.svelte
